Create authentication route from server
