+++ server.js
@@ -96,6 +96,13 @@ app.post("/upload", upload.single("image"), async (req, res) => {
 });
 
 // Routes
+
+// Import routes
+const authRoutes = require("./routes/auth");
+
+// Use routes
+app.use(authRoutes);
+
 app.get("/", (req, res) => {
   res.render("index", {
     title: "Nature Nexus - Home",
@@ -118,14 +125,6 @@ app.get("/explore", (req, res) => {
     error: null,
     currentPage: "explore",
     user: req.session.user || null,
-  });
-});
-
-app.get("/login", (req, res) => {
-  res.render("login", {
-    title: "Nature Nexus - Login",
-    error: null,
-    currentPage: "login",
   });
 });
 
@@ -168,84 +167,6 @@ app.get("/tiles/:z/:x/:y", async (req, res) => {
   }
 });
 
-app.post("/login", async (req, res) => {
-  const { email, password } = req.body;
-  try {
-    const user = await User.findOne({ email });
-    if (user && (await bcrypt.compare(password, user.password))) {
-      req.session.user = user;
-      res.redirect("/");
-    } else {
-      res.render("login", {
-        title: "Nature Nexus - Login",
-        error: "Invalid email or password",
-        currentPage: "login",
-      });
-    }
-  } catch (err) {
-    console.error("Login error:", err);
-    res.render("login", {
-      title: "Nature Nexus - Login",
-      error: "An error occurred",
-      currentPage: "login",
-    });
-  }
-});
-
-app.get("/register", (req, res) => {
-  res.render("login", {
-    title: "Nature Nexus - Register",
-    error: null,
-    currentPage: "register",
-  });
-});
-
-app.post("/register", async (req, res) => {
-  const { email, password } = req.body;
-  console.log("Registration attempt:", { email, password: "***" });
-  if (!email || !password) {
-    console.log("Missing email or password");
-    return res.render("login", {
-      title: "Nature Nexus - Register",
-      error: "Email and password are required",
-      currentPage: "register",
-    });
-  }
-  try {
-    const existingUser = await User.findOne({ email });
-    if (existingUser) {
-      console.log("Email already registered:", email);
-      res.render("login", {
-        title: "Nature Nexus - Register",
-        error: "Email already registered",
-        currentPage: "register",
-      });
-    } else {
-      const saltRounds = process.env.SALT_ROUND
-        ? parseInt(process.env.SALT_ROUND)
-        : 10;
-      const hashedPassword = await bcrypt.hash(password, saltRounds);
-      console.log("Password hashed successfully");
-      const newUser = new User({
-        email,
-        password: hashedPassword,
-        stats: { birds: 0, plants: 0 },
-        collections: [],
-      });
-      const savedUser = await newUser.save();
-      console.log("User saved to MongoDB:", savedUser._id);
-      res.redirect("/login");
-    }
-  } catch (err) {
-    console.error("Registration error:", err);
-    res.render("login", {
-      title: "Nature Nexus - Register",
-      error: `Registration failed: ${err.message}`,
-      currentPage: "register",
-    });
-  }
-});
-
 app.get("/collections", (req, res) => {
   res.render("collection", {
     title: "Nature Nexus - Collections",
@@ -264,12 +185,6 @@ app.get("/collection", (req, res) => {
   } else {
     res.redirect("/login");
   }
-});
-
-app.get("/logout", (req, res) => {
-  req.session.destroy(() => {
-    res.redirect("/login");
-  });
 });
 
 // Added identify route
@@ -643,11 +558,9 @@ app.post("/submitSighting", async (req, res) => {
       res.status(500).json({ error: "Failed to save sighting" });
     }
   } else {
-    return res
-      .status(404)
-      .json({
-        error: "User not found. You must be logged in to submit a sighting.",
-      });
+    return res.status(404).json({
+      error: "User not found. You must be logged in to submit a sighting.",
+    });
   }
 });
